+++ app/src/main/java/com/example/pilotlogbook/data/room/entities/dailyflight/tuples/ArrivalTuple.kt
@@ -4,5 +4,5 @@ import androidx.room.ColumnInfo
 
 data class ArrivalTuple(
     @ColumnInfo(name = "arrival_place")val place: String,
-    @ColumnInfo(name = "arrival_time")val time: Int?
+    @ColumnInfo(name = "arrival_time")val time: Long?
 )
